2D Next Fit Bin Packing
Started 2D Bin Packing.

Bins are trees with each node having an x and y
capacity and an occupied bool. Tree nodes have
left and bottom children.

Items are namedtuples with x and y values. When
Items are inserted into a bin tree, insert()
recurses over child nodes until it finds the first
unoccupied node that fits the Item. The node is
resized to Item and two child nodes are created
to take up the remaining space to the right and
below the item.

Finished:
-Basic tree functionality is complete.

To Do:
-Modified AVL Tree for sorting BinTrees.
-BinTree traversal method which returns the bin
  layout in a renderable format.
-Wrapper function to handle generating new Bins,
  and inserting items into BinTree roots.